Improve error message when file not found
The current logic could confuse a user who enters a filepath to
a file that does not exist. The current logic states, including
when the file does not exist, that the document can't be parsed.
This could imply to the user that they typed the filename correctly
and the file does exist. The new logic alerts a user to the fact
that the file does not exist.

Signed-off-by: John Speed Meyers <jsmeyers@chainguard.dev>